recreated folder structure
Recreated folder structure from the old version of the ARGP, complete with having copied over all of the contents of each of the folders.

Everything folder and file wise should now exist in the folder to run the experiment, except for the qualisys folders. I'll include that in next commit