trentwirth_dev1 (#13)
* recreated folder structure

Recreated folder structure from the old version of the ARGP, complete with having copied over all of the contents of each of the folders.

Everything folder and file wise should now exist in the folder to run the experiment, except for the qualisys folders. I'll include that in next commit

* added qualisys SDK to project

they claim they've only tested it up to Unity 2019.4.23f1.

this could pose a problem later... keeping an eye on things.

* added unity objects

added all objects to scene except for UXF_Rig and EventSystem

* connected scripts to game objects

began the process of reconnecting scripts to game objects. Ran into an issue (that I've had before) where the GenerateTerrain method of the class LoadTerrain didn't show up in the UXF rig. Need to investigate this problem further.