Trentwirth/in lab dev scripts and game objects (#19)
* connected scripts to game objects

added scripts to game objects that interact with the UXF Event System. Including: start session, start trial, end trial. Connected LoadTerrain.GenerateTerrain to the UXF Rig via the gameObject `ScriptHolder`.

Removed the distractor, and changed the target to a bullseye. Connected audio trigger script to the bullseye target, and set the dwell-time-threshold to 20 (i.e., what it was set at before, established via testing with Michael)

* targets & bad bar-logic

targets are being built, hooray! But, the start and end bar logic is not working - bad.

Goal for near future: attach scripts to game-objects with an "Initialization" script, so I can get away from drag and drop.

* cleaned up code w/ Jon

* refactored LoadTerrain.cs