Trentwirth/generate start end bars (#24)
* generate start & end bars

added script that generates start and end bars based on the position of the room the participant starts in

* cleaned up start/end bar generation

added an if statement that checks to see if a start bar exists, if it doesn't, the script creates a start and end bar. This was done to avoid creating multiple start/end bar pairs for multiple marker collisions

* created a timer

the current timer only generates a sound if the trial isn't completed in 12 (parameterized number) seconds. I *should* create a better timer that uses Coroutines (a unity feature that I didn't want to mess with, but think I should).

* reverting this

* Revert "reverting this"

This reverts commit c23cdafaa5df88a99617b2d5646417af9b99efe5.

* Revert "created a timer"

This reverts commit 3424b8f24f4aeaf4f81b6b4743214d054b4b871b.