all targets - generate terrain files
created an all targets version of the experiment, and successfully can load in the new trial list. Important: I didn't re-write the matlab code used to generate the terrain, but I decided to continue with `.csv` files because that's how I currently attach the target objects to the projectors. When we implement the "flash light" method that Michael is working on, I can also take that as an opportunity to (by necessity) update how I generate terrain.